Updated files with authentication and db reads
