+++ frontend/components/HomeScreen.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { StyleSheet, Text, View, TouchableOpacity, Image, FlatList } from 'react-native';
+import { StyleSheet, Text, View, TouchableOpacity, ScrollView, Image, FlatList } from 'react-native';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 import { db } from '../firebase/config.js';
 import { ref, onValue } from 'firebase/database';
@@ -99,6 +99,7 @@ const HomeScreen = ({navigation}) => {
         </View>
       </View>
 
+      <ScrollView style={styles.overallContainer}>
       <View style={styles.section}>
         <View style={styles.sectionHeader}>
           <Text style={styles.sectionTitle}>Suggested for you</Text>
@@ -128,11 +129,16 @@ const HomeScreen = ({navigation}) => {
           contentContainerStyle={styles.flatListContainer}
         />
       </View>
+      </ScrollView>
     </View>
   );
 };
 
 const styles = StyleSheet.create({
+  overallContainer: {
+    flex: 1,
+    backgroundColor: "#fff",
+  },
   container: {
     flex: 1,
     paddingTop: 20,
