+++ frontend/components/MyReferralsScreen.js
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { View, Text, StyleSheet } from 'react-native';
 
-const MyRewardsScreen = () => (
+const MyReferralsScreen = () => (
   <View style={styles.container}>
-    <Text>My Rewards Screen</Text>
+    <Text>My Referrals Screen</Text>
   </View>
 );
 
@@ -15,4 +15,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default MyRewardsScreen;
+export default MyReferralsScreen;
